Bug #1522916 staging: sprd: sipa: solve coverity warning
[root cause  ] coverity warning
[changes     ] sipa
[side effects] na
[self test   ]: yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: ums9620-2h10
[test case]: sipa
[reviewers   ] qingsheng.li
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Solve coverity

+++ drivers/staging/sprd/sipa_v3/sipa_phy_r3p0/sipa_glb_phy.c
@@ -1135,8 +1135,8 @@ static int ipa_phy_monitor_ipa_or_tft(void __iomem *reg_base, u32 flag)
 	tmp |= (flag << 14);
 	writel_relaxed(tmp, reg_base + IPA_MODE_N_FLOWCTRL);
 
-	if ((readl_relaxed(reg_base + IPA_MODE_N_FLOWCTRL) &
-	     IPA_MONITOR_IPA_OR_TFT_MASK) != (flag << 14))
+	if ((u32)(readl_relaxed(reg_base + IPA_MODE_N_FLOWCTRL) &
+		  IPA_MONITOR_IPA_OR_TFT_MASK) != (flag << 14))
 		return -EIO;
 
 	return 0;
